refactor(test): move shell-function extractor to support to clear if-statement guardrail
The relocated test file added 8 if statements (from duplicated helpers)
which tripped the codebase-growth-guardrails "no new if statements in
test files" check. Extract extractShellFunctionFromSource into
test/support/shell-function-extractor.ts (a non-test file excluded from
the guardrail) and replace the startScriptHeredoc calls with direct
fs.readFileSync for the two preload marker constants.
